+++ datacrunch/http_client/http_client.py
@@ -33,7 +33,7 @@ class HTTPClient:
         self._auth_service.authenticate()
 
     def post(self, url: str, json: dict = None, **kwargs) -> requests.Response:
-        """Sends a POST request. 
+        """Sends a POST request.
 
         A wrapper for the requests.post method.
 
@@ -59,8 +59,35 @@ class HTTPClient:
 
         return response
 
+    def put(self, url: str, json: dict = None, **kwargs) -> requests.Response:
+        """Sends a PUT request.
+
+        A wrapper for the requests.put method.
+
+        Builds the url, uses custom headers, refresh tokens if needed.
+
+        :param url: relative url of the API endpoint
+        :type url: str
+        :param json: A JSON serializable Python object to send in the body of the Request, defaults to None
+        :type json: dict, optional
+
+        :raises APIException: an api exception with message and error type code
+
+        :return: Response object
+        :rtype: requests.Response
+        """
+        url = self._add_base_url(url)
+        headers = self._generate_headers()
+
+        self._refresh_token_if_expired()
+
+        response = requests.put(url, json=json, headers=headers, **kwargs)
+        handle_error(response)
+
+        return response
+
     def get(self, url: str, params: dict = None, **kwargs) -> requests.Response:
-        """Sends a GET request. 
+        """Sends a GET request.
 
         A wrapper for the requests.get method.
 
@@ -68,7 +95,7 @@ class HTTPClient:
 
         :param url: relative url of the API endpoint
         :type url: str
-        :param params: Dictionary, list of tuples or bytes to send in the query string for the Request., defaults to None
+        :param params: Dictionary, list of tuples or bytes to send in the query string for the Request. defaults to None
         :type params: dict, optional
 
         :raises APIException: an api exception with message and error type code
@@ -87,7 +114,7 @@ class HTTPClient:
         return response
 
     def delete(self, url: str, json: dict = None, **kwargs) -> requests.Response:
-        """Sends a DELETE request. 
+        """Sends a DELETE request.
 
         A wrapper for the requests.delete method.
 
@@ -124,7 +151,7 @@ class HTTPClient:
             # to to refresh. if refresh token has expired, reauthenticate
             try:
                 self._auth_service.refresh()
-            except:
+            except Exception:
                 self._auth_service.authenticate()
 
     def _generate_headers(self) -> dict:
@@ -141,7 +168,7 @@ class HTTPClient:
         return headers
 
     def _generate_bearer_header(self) -> str:
-        """generate the authorization header Bearer string 
+        """generate the authorization header Bearer string
 
         :return: Authorization header Bearer string
         :rtype: str
@@ -162,7 +189,7 @@ class HTTPClient:
     def _add_base_url(self, url: str) -> str:
         """Adds the base url to the relative url
 
-        example: 
+        example:
         if the relative url is '/balance'
         and the base url is 'https://api.datacrunch.io/v1'
         then this method will return 'https://api.datacrunch.io/v1/balance'