+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+v1.0.0 (2021-06-16)
+-------------------
+
+* Added support for storage volumes
+* Breaking change: moved all constants under DataCrunchClient to DataCrunchClient.constants
+
 v0.1.0 (2021-01-05)
 -------------------